+++ src/extension/grpc/parser/h2/types.ts
@@ -1,0 +1,1 @@
+export * from '@buf/stateful_runme.bufbuild_es/runme/parser/v1/parser_pb'
