+++ src/extension/serializer.ts
@@ -22,22 +22,6 @@ import {
 import { ulid } from 'ulidx'
 import { maskString } from 'data-guardian'
 import YAML from 'yaml'
-import { ParserService } from '@buf/stateful_runme.connectrpc_es/runme/parser/v1/parser_connect'
-import { createGrpcTransport, GrpcTransportOptions } from '@connectrpc/connect-node'
-import { createClient as createConnectClient, Client as ConnectClient } from '@connectrpc/connect'
-// ts bindings generated by protoc-gen-es
-import {
-  RunmeIdentity,
-  RunmeSession,
-  Notebook,
-  Frontmatter,
-  CellOutput,
-  CellKind,
-  CellExecutionSummary,
-  DeserializeRequest,
-  SerializeRequest,
-  SerializeRequestOptions,
-} from '@buf/stateful_runme.bufbuild_es/runme/parser/v1/parser_pb'
 
 import { Serializer } from '../types'
 import {
@@ -51,7 +35,27 @@ import {
 } from '../constants'
 import { ServerLifecycleIdentity, getSessionOutputs, getTLSEnabled } from '../utils/configuration'
 
-import { type ReadyPromise } from './grpc/client'
+import {
+  ParserService,
+  createGrpcTransport,
+  GrpcTransportOptions,
+  createConnectClient,
+  ConnectClient,
+} from './grpc/parser/h2/client'
+// ts bindings generated by protoc-gen-es
+import {
+  RunmeIdentity,
+  RunmeSession,
+  Notebook,
+  Frontmatter,
+  CellOutput,
+  CellKind,
+  CellExecutionSummary,
+  DeserializeRequest,
+  SerializeRequest,
+  SerializeRequestOptions,
+} from './grpc/parser/h2/types'
+import { type ReadyPromise } from './grpc/tcpClient'
 import Languages from './languages'
 import { PLATFORM_OS } from './constants'
 import { initWasm } from './utils'
