+++ tests/extension/extension.test.ts
@@ -26,7 +26,7 @@ vi.mock('../../src/extension/panels/panel', () => {
   return { default: Panel }
 })
 
-vi.mock('../../src/extension/grpc/client', () => {
+vi.mock('../../src/extension/grpc/tcpClient', () => {
   class MockedParserServiceClient {
     deserialize = vi.fn(() => {
       return { status: { code: 'OK' } }
