+++ tests/extension/provider/notebook.test.ts
@@ -8,7 +8,7 @@ import { StatefulAuthProvider } from '../../../src/extension/provider/statefulAu
 vi.mock('vscode')
 vi.mock('vscode-telemetry')
 
-vi.mock('../../../src/extension/grpc/client', () => ({
+vi.mock('../../../src/extension/grpc/tcpClient', () => ({
   ParserServiceClient: vi.fn(),
 }))
